Create Number Guessing Game
Number Guessing Game (Java)

A simple Java console-based game where the computer randomly generates a number between 0 and 99, and the user tries to guess it. The program gives hints whether the guessed number is large or small until the correct number is guessed (or the user quits by entering a negative number).